refactor(DC-704): extract configuration sources into ConfigurationSetup
